Dispatch update Actions(EntryDetail-reset) in Redux in React Native
Made the rest button to reset the information for a specific dat when clicked

Why we use shouldComponentUpdate explanation below
When we click the Reset Button, this Component EntryDetail will re-render because it is still in the Route Stack even though we are going back to home. This will make our app to break because our Component will pass the entryId to MetricCard Component which will try to get a state at the specific ID entry passed as prop by EntryDetail which we just reset. So it will be undefined.
Undefined variables makes your app crash

***We use the shouldComponentUpdate to allow the EntryDetails to re-render on only when it contains the metrics property and the today property inside the metrics is undefined

So the EntryDetail Component will not re-render if the data Entry data is null or contains the 'today' property that says 'Don't forget to log your data today!'