First Service Permission in React Native
We style the UI for if the service is undetermined.
This scenario is where user decides to neither deny nor grant you permission to their location.

In an ideal world, the user would always grant you permission to whatever you'd like, but, this isn't always the case and as a UI developer, you need to plan accordingly for those moments.