Add Animation to directional changes in Live
We want to animate the text that says the direction your are going on our Live component. We want to bounce the text

Our bounceValue starts at 1. When the location changes, we will take that value from 1, time it taking 200milliseconds up to 1.04
An then take it back down to 1 using a spring animation

NOTE: we tied the bounceValue to ur UI
I use the useRef Hook to return a mutable ref object. This ref object's current property is initialized as the given argument and persists throughout the component lifecycle.
https://docs.expo.io/versions/v40.0.0/react-native/animated/

Animated provides three types of animation configurations: timing, spring, and decay.

To eliminate the waring
https://hassanannajjar.medium.com/how-to-fix-animated-usenativedriver-was-not-specified-fb7807b1efc0