Implemented Local Notifications in React Native
In our App everyday at 8am, if the user have not logged their entry for that day, we will send them a notification reminder to log their information for that day.
In the Lab we did
https://github.com/udacity/reactnd-UdaciFitness-complete/blob/setLocalNotification/utils/helpers.js
https://github.com/udacity/reactnd-UdaciFitness-complete/commit/63778456f674355e40044c673f4b966ebd446866
- The idea is that we have a key in the AsyncStorage that could be true or false. If it is true, it means that we already set a notification
- we set the key in the async storage to false by always deleting the key and setting up a new notification at the AddEntry Component each time the user submits an entry.
- We set the try to make sure that the user have already has a scheduled notification every time the App component mounts!

However that version is outdated and i did not full implement this for this project. However check out
https://docs.expo.io/versions/v40.0.0/sdk/notifications/

NOTE:
Local notifications do not use or require any external infrastructure; they happen entirely on the device itself. That means that the only requirement for the device to display the notification is that the device is on. On the other hand, push notifications require you to have a server which handles pushing the notification to your user's devices when a certain event occurs.

But since we're not incorporating an external server, and all the logic about when we should show the notification can be done on the phone itself -- local notifications will be the most ideal for our application.

IMPORTANT:
To schedule a local notification to be shown in the future you use Expo's Notifications.scheduleNotificationAsync method.

You can exploer sending notificatios to users that subscribe to a specific chanel an all that on your own with expo